chore: rename directory helper/ to util/
