+++ pkg/repo/gitlab/gitlab_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/nikoksr/proji/pkg/helper"
+	"github.com/nikoksr/proji/pkg/util"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -35,7 +35,7 @@ var apiToken = os.Getenv("PROJI_AUTH_GL_TOKEN")
 
 // TestNew tests the creation of a new github object based on given github URLs.
 func TestNew(t *testing.T) {
-	helper.SkipNetworkBasedTests(t)
+	util.SkipNetworkBasedTests(t)
 
 	// These should work
 	for i, repo := range goodRepos {
@@ -58,7 +58,7 @@ func TestNew(t *testing.T) {
 // TestGitLab_LoadTreeEntries tests the github method TestGetTreePathsAndTypes which tries
 // to request and receive the folders paths and types of a github repo tree.
 func TestGitLab_LoadTreeEntries(t *testing.T) {
-	helper.SkipNetworkBasedTests(t)
+	util.SkipNetworkBasedTests(t)
 
 	type testEntry struct {
 		id        string
@@ -122,12 +122,12 @@ func TestGitLab_LoadTreeEntries(t *testing.T) {
 				},
 				{
 					id:        "e69de29bb2d1d6434b8b29ae775ad8c2e48c5391",
-					path:      "include/helper.hpp",
+					path:      "include/util.hpp",
 					entryType: "blob",
 				},
 				{
 					id:        "e69de29bb2d1d6434b8b29ae775ad8c2e48c5391",
-					path:      "src/helper.cpp",
+					path:      "src/util.cpp",
 					entryType: "blob",
 				},
 				{
@@ -193,7 +193,7 @@ func TestGitLab_LoadTreeEntries(t *testing.T) {
 				},
 				{
 					id:        "e69de29bb2d1d6434b8b29ae775ad8c2e48c5391",
-					path:      "include/helper.hpp",
+					path:      "include/util.hpp",
 					entryType: "blob",
 				},
 				{
@@ -203,7 +203,7 @@ func TestGitLab_LoadTreeEntries(t *testing.T) {
 				},
 				{
 					id:        "e69de29bb2d1d6434b8b29ae775ad8c2e48c5391",
-					path:      "src/helper.cpp",
+					path:      "src/util.cpp",
 					entryType: "blob",
 				},
 				{
