+++ pkg/util/util.go
@@ -1,4 +1,4 @@
-package helper
+package util
 
 import (
 	"fmt"
