+++ pkg/util/util_test.go
@@ -1,4 +1,4 @@
-package helper
+package util
 
 import (
 	"os"
@@ -14,7 +14,7 @@ func TestDoesPathExist(t *testing.T) {
 		path   string
 		exists bool
 	}{
-		{path: "./helper_test.go", exists: true},
+		{path: "./util_test.go", exists: true},
 		{path: "../../README.md", exists: true},
 		{path: "../../READMENOT.md", exists: false},
 		{path: "./CrYpTicFiLe.txt", exists: false},
